Scan Photostream photos for text and remove (#192)
Added an Android Native Module that loads in a photo, uses Google's ML Kit to scan the photo for text, and blurs out the text if any is found. The native module interfaces with RN entirely within PhotostreamImageSelectionView.tsx.

All the heavy lifting is in ImageTextBlur.kt. This change does add a new library to the app, which may increase the app's size by ~4MB.

A similar native module could be made for iOS, but it doesn't look like the iOS build is working.

Co-authored-by: Chall Fry <chall@challfry.com>